http://www.atutor.ca/atutor/mantis/view.php?id=2449

+++ docs/mods/_standard/links/module_groups.php
@@ -13,16 +13,14 @@ function links_create_group($group_id) {
 function links_delete_group($group_id) {
 	global $db;
 
-	$sql	= "SELECT * FROM ".TABLE_PREFIX."links_categories WHERE owner_id=$group_id";
+	$sql	= "SELECT cat_id FROM ".TABLE_PREFIX."links_categories WHERE AND owner_type=".LINK_CAT_GROUP." owner_id=$group_id";
 	$result = mysql_query($sql, $db);
 
-	$total_links = 0;
-	while ($row = mysql_fetch_array($result)) {
-		$sql = "DELETE FROM ".TABLE_PREFIX."links WHERE cat_id=$row[0]";
+	while ($row = mysql_fetch_assoc($result)) {
+		$sql = "DELETE FROM ".TABLE_PREFIX."links WHERE cat_id=$row[cat_id]";
 		$result2 = mysql_query($sql, $db);
-		$total_links += mysql_affected_rows($db);
 	}
-	$sql	= "DELETE FROM ".TABLE_PREFIX."links_categories WHERE owner_id=$group_id";
+	$sql	= "DELETE FROM ".TABLE_PREFIX."links_categories WHERE owner_type=".LINK_CAT_GROUP." AND owner_id=$group_id";
 	$result = mysql_query($sql, $db);
 }
